Final Commit
Finished comments on style sheet and completely finished working on the style sheet. Throughout the entire process the website has gone through many changes to become more aesthetically pleasing and just an overall better looking website. The last two steps ended with me editing the names of the original authors and the directors so that they become bolder and more eye-catching.